added a suggestion section
The name of this section can be changed, but I thought that it would useful to have some more detailed instructions on common issues. I added a word on how to fix line ending problems. This happened to me twice, and it is difficult to spot the problem. More suggestions should be added when we encounter other common problems.